# Style-profiling/mlp-data-service/schemas/static/openapi/api_spec.yaml
openapi: 3.0.1
info:
  title: MLP Data Service
  description: MLPlatform Data Management Service Backend
  version: 0.4.1
servers:
  - url: http://localhost:9890
...
paths:
...
  /api/v1/datasets/{id}:
    get:
      summary: ''
      tags:
        - Dataset
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': 
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
    patch:
      summary: ''
      tags:
        - Dataset
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetUpdateDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: ''
      description: ''
      tags:
        - Dataset
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
...
  /api/v1/versioned-datasets/{id}:
    get:
      summary: ''
      tags:
        - Versioned Dataset
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': 
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionedDataset'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: ''
      description: ''
      tags:
        - Versioned Dataset
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
...
  /api/v1/experiments/{id}:
    get:
      summary: ''
      tags:
        - Experiment
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': 
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
    patch:
      summary: ''
      description: 'Updating is only possible when the status is created.'
      tags:
        - Experiment
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentUpdateDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: ''
      description: 'Deletion is possible only when the status is created, done or failed.'
      tags:
        - Experiment
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        default:
          $ref: '#/components/responses/UnknownError'
...
components:
  schemas:
    Algorithm:
      type: object
      properties:
        name:
          type: string
        image_name:
          type: string
        training_params:
          type: array
          items:
            type: object
            required: [name, type, default, ]
            properties:
              name:
                type: string
              type:
                type: string
              argument:
                type: string
              default:
                type: string
        performance_metrics:
          type: array
          items:
            type: object
            required: [name, type, ]
            properties:
              name:
                type: string
              type:
                type: string
              value:
                type: string
    Attribute:
      type: object
      required: [key, value, ]
      properties:
        key:
          type: string
        values:
          type: array
          items:
            type: string
...
    Dataset:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        description: 
          type: string
        ml_task:
          type: integer
          description: |
            ml_task
            - 0: text deep tagging
            - 1: image deep tagging
            - 2: virtual try-on
          enum: [0, 1, 2]
        cover_image:
          type: string
        labeled_data:
          type: integer
          description: 'Annotated Data의 개수'
          minimum: 0
        total_data:
          type: integer
          description: '총 Data의 개수'
          minimum: 0
        history:
          type: object
          description: '표준필드 : 로그 데이터'
        ext:
          type: object
          description: '표준필드 : 확장 데이터'
        reg_date:
          type: string
          description: '표준필드 : 등록일'
          format: date-time
        reg_id:
          type: string
          description: '표준필드 : 등록ID'
        mod_date:
          type: string
          description: '표준필드 : 수정일'
          format: date-time
        mod_id:
          type: string
          description: '표준필드 : 수정ID'
...
    VersionedDataset:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        ml_task:
          type: integer
          description: |
            ml_task
            - 0: text deep tagging
            - 1: image deep tagging
            - 2: virtual try-on
          enum: [0, 1, 2]
        dataset_name:
          type: string
        index:
          type: integer
          minimum: 0
        name:
          type: string
        gen_date:
          type: string
          description: '파일 생성일'
          format: date-time
        split:
          type: array
          description: 'train, valid, test 개수'
          items:
            type: integer
          minItems: 3
          maxItems: 3
        status:
          type: integer
          description: |
            status
            - 0: queued
            - 1: processing
            - 2: done
            - 3: failed
          enum: [0, 1, 2, 3]
        attribute_info:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
        history:
          type: object
          description: '표준필드 : 로그 데이터'
        ext:
          type: object
          description: '표준필드 : 확장 데이터'
        reg_date:
          type: string
          description: '표준필드 : 등록일'
          format: date-time
        reg_id:
          type: string
          description: '표준필드 : 등록ID'
        mod_date:
          type: string
          description: '표준필드 : 수정일'
          format: date-time
        mod_id:
          type: string
          description: '표준필드 : 수정ID'
...
    Experiment:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
        description:
          type: string
        ml_task:
          type: integer
          description: |
            ml_task
            - 0: text deep tagging
            - 1: image deep tagging
            - 2: virtual try-on
          enum: [0, 1, 2]
        dataset_name:
          type: string
        versioned_dataset_name:
          type: string
        versioned_dataset_index:
          type: integer
        dataset_id:
          type: string
        model_name:
          type: string
        algorithms:
          type: array
          items:
            $ref: '#/components/schemas/Algorithm'
        artifacts:
          type: array
          items:
            type: object
        status:
          type: integer
          description: |
            status
            - 0: created
            - 1: queued
            - 2: training
            - 3: done
            - 4: failed
          enum: [0, 1, 2, 3, 4]
        history:
          type: object
          description: '표준필드 : 로그 데이터'
        ext:
          type: object
          description: '표준필드 : 확장 데이터'
        reg_date:
          type: string
          description: '표준필드 : 등록일'
          format: date-time
        reg_id:
          type: string
          description: '표준필드 : 등록ID'
        mod_date:
          type: string
          description: '표준필드 : 수정일'
          format: date-time
        mod_id:
          type: string
          description: '표준필드 : 수정ID'
...
    ObjectId:
      type: string
      description: id 
    ResponseErrorDto:
      type: object
      required: [error_code, error_message]
      properties:
        error_code:
          type: integer
          description: ''
        error_message:
          type: string
          description: ''
...
    DatasetUpdateDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        cover_image:
          type: string
...
    ExperimentUpdateDto:
      type: object
      properties:
        name:
          type: string
        algorithms:
          type: array
          items:
            $ref: '#/components/schemas/Algorithm'
        status:
          type: integer
          description: |
            status
            - 0: created
            - 1: queued
            - 2: training
            - 3: done
            - 4: failed
          enum: [0, 1, 2, 3, 4]
  parameters:
...
    Id:
      name: id
      in: path
      description: id
      required: true
      schema:
        $ref: '#/components/schemas/ObjectId'
  responses:
    UnknownError:
      description: unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseErrorDto'
    204NoContent:
      description: No Content
    400BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseErrorDto'
    404NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseErrorDto'
...